pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ibkr-eda"
version = "0.1.0"
description = "Exploratory data analysis toolkit for Interactive Brokers via Client Portal API"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "requests>=2.31,<3",
    "pandas>=2.1,<3",
    "python-dotenv>=1.0,<2",
    "urllib3>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "responses>=0.25",
]
dashboard = [
    "flask>=3.0",
]

[tool.setuptools.packages.find]
include = ["ibkr_eda*"]
